Merged PR 1693: Fixed bug in webhook external exception handler
The bug would only occur if there was a socket-type exception (like timeout) when calling a webhook that came from an intent with no variables. That's because the intent would get fulfilled as soon as it was triggered, so the `currentIntent` variable would still be null by the time we are handling a webhook exception. (See ChatLogic.java for changes related only to the fix)

I have also taken the opportunity to refactor some tests, add a couple and fix a few tests that were throwing expected exceptions for the wrong reasons.

Related work items: #3100