Merged PR 1740: Added support for bot cloning in the API.
Added support for bot cloning in the API.
CloneAI uses a similar interface as CreateAI since it's meant (at least for now) to be used by the same place on the UI. Basically it supports adding overrides to the bot to be cloned so that on the creation screen the user can tweak parameters like name, description, voice, etc. These overrides will then be used instead of the ones from the original AI.
Also, it will create a new name (add a random suffix) if it detects taht somehow the cloning will end up with the same name as the original, since we don't support duplicate AI names.

We should eventually move from passing a form to using a JSON for the AI structure (both create and clone overrides) but that will require UI changes so keeping the same structure for now.

Related work items: #3164