Merged PR 1744: Make an AI read-only when it's sent for publication.
Make an AI read-only when it's sent for publication.
An AI becomes RO when it's publishing state is SUBMITTED, PUBLISHED or REMOVED.
The following operations are then not allowed when the AI is RO:
- UpdateAI
- SetAiBotConfigDescription
- SetAiBotConfig
- DeleteAI
- LinkBotToAI
- UnlinkBotFromAI
- WriteIntent
- DeleteIntent
- UploadTrainingFile
- StartTraining
- StopTraining
- UpdateTraining

Not included but open for consideration are:
- Facebook link/unlink

Fixed a couple of UTs which were not reaching the expected code to test.

Related work items: #3169