Merged PR 1752: Added a POST /ai/{aiid}/bots endpoint to allow direct updating of the linked
Added a POST /ai/{aiid}/bots endpoint to allow direct updating of the linked bots.
This will replace the existing linked bot list.
For instance, if the user has linked bots {1, 2, 3} and we call this endpoint with {1, 3, 5}, the latter will be the list that will remain (assuming the bots are valid and owned). The API will create a diff between the previous and new states and only apply the changes required; for instance if the list passed is equal to the exising list nothong will happen.
The changes are wrapped around a transaction, so the update is guaranteed to be atomic (either succeeds with all the updates or fails and no update is performed).

Related work items: #3167