Merged PR 1774: Enhanced the UT that tested the max number of linked skills to include more
Enhanced the UT that tested the max number of linked skills to include more already linked bots and to also test the logic of linking after hitting the limit and unlinking one.

Note that this doesn't fix #3245, it only proves that the API is correctly testing the limitation. The fix for that bug will lie on the UI (or moving to the new endpoint that bulk-updates the links within a transaction)

Related work items: #3245