Merged PR 1808: Make heartbeat calls in separate threads
Regardless of the fact that we set short timeouts on http calls there are edge cases where the timeouts are ignored and it can take 15 minutes (!) for a DNS call to fail. Isn't Java wonderful?
The solution is to make the heartbeat http call in a separate thread so that we can wait exactly _n_ milliseconds before bailing out and declaring the call a fail.

Related work items: #2707