Merged PR 1812: More consistent (and complete) validation for intents
Add more validation for intents at API level to avoid things going wrong when we try to write them to the database.
This changes intent create and edit, as well as bot import.
The front-end was already validating most of these things so there should be few changes to the end-user.

Related work items: #3342