Merged PR 1817: Speed up recovery of thread pool by not waiting too long for http calls to complete
We currently wait for the default time, which is very long, for an http call to the back-end to complete. The call is threaded so at a certain point (20s) the chat logic will give up and treat the http call as failed - however, the worker thread waits for the call to complete before cleaning up. This means we can run out of threads when latency goes up on one or more of the back-end servers.

This change puts a hard limit on the read and connect timeouts for the http call, so that they timeout and free up the thread after the chat logic's maximum wait time has elapsed.

Related work items: #3353