Merged PR 1848: Fixed intents not being saved
Fixed intents not being saved
We were using a try-with-resources block in writeIntent for the transaction, and we weren't commiting it since this is part of a bigger operation (transaction passed != null). But exiting the try-with-resources without committing would effectively roll it back. So we need just to use a try/catch block and close manually, if required.
Also, the main block in writeIntent didn't commit the transaction, so that would have not saved it in any case.

Related work items: #3427