Merged PR 1920: Fix for NPE during loading an intent.
Fix for NPE during loading an intent.
For some reason (at least) one user hit the case where the chat references an intent that cannot be loaded from the db.
This fix logs a warning so we can see if this happens more often, and lets the chat go through without using that memory intent (since it doesn't seem to exist anyway).
We will need to monitor this to determine the root cause since it's not clear how it got into this state.

Also passed the DEVID to this method to allow later aggregating the logs per user.

Related work items: #3495