Merged PR 1918: Facebook integration: ignore failure to clear customisations on disconnect and delete
We are seeing users on prod who cannot disconnect a bot from Facebook because of an exception from the Facebook API when we are trying to clear the customisations.
The error says `OAuthException 100: (#100) You must set a Get Started button if you also wish to use persistent menu.`, which doesn't make much sense to us because we don't set persistent menus. All we need to do is treat this as a warning and carry on with the disconnect or delete operation.

Related work items: #3181