Merged PR 1919: Created client/server (local) for controller service.
Created client/server (local) for controller service.
This tuns still within the core-api tomcat process, but now all controller-related operations are requested through JAX-RS endpoints.
Move the queue processor to the controller-service package since this is meant to be run there. Core-services just puts items in the queue but
it's controller-service that processes the queue items.

There's most likely some code simplification to be done but for now it's more important to get the service split and stable.

Related work items: #3466, #3467