Merged PR 1931: Catch a webhook exception to allow the intent to be cleared in case of error.
Catch a webhook exception to allow the intent to be cleared in case of error.
Because we were just bubbling up the WebHookException, when a webhook failed it meant that the state of the last variable wouldn't have the chance to be set, and the intent wouldn't be cleared, leaving it in a weird state. Then triggering the intent again would result in all but the last entities to be already fulfilled.
In the long run we may want to define a better strategy for handling webhook failures (retry policies, used defined config, etc) but for now just clear the intent so that it provides a consistent experience.

Related work items: #3556