Merged PR 1962: Fix exception thrown when trying to pass chat with curly brackets to back-end servers
If we enter the text "{hello}" into the chat window we get an ISE because the JerseyWebTarget interprets the curly braces as templating placeholders.
To fix this we use templating placeholders and _resolve_ them with the text we want to pass.
Also, once we're at it I thought the chat coming from Facebook really needed to pass through the same validation routine as chat from the API.

Related work items: #3553