Merged PR 2022: Fix for importing a bot without a training file.
Fix for importing a bot without a training file.
On AILogic::createImportedBot(), one of the last steps of importing was to upload the training file. When this is null (not present in the JSON), it was still trying to write it to database, which would fail and cause the "Failed to add training file". Simple fix was just to only update the db when needed.

Related work items: #3667