Merged PR 2026: Fix for import bot operation failing due to transactional issues.
Fix for import bot operation failing due to transactional issues.
Some of the operations were using try-with-resources for the transaction which would complete (commit) the transaction at the end of the block. Then subsequent calls would be run outside the context of that transaction, causinng the failure in correctly importing the data.

Related work items: #3625, #3636