Merged PR 2093: Added endpoint for regenerating tokens for users (single/multiple).
Added endpoint for regenerating tokens for users (single/multiple).
The /admin/regenerate_tokens will traverse all the users table and check if the developer JWT token can be correctly decoded. For users which the roken is incorrect, it will re-generate a new one, using the plan id as the base for the security role.
The /admin/{devId}/regenerate_token will regenerate the token for a user (if needed).
Passing the query parameter "dryrun=1" will run the code but not apply changes to the database.

Related work items: #3812