Merged PR 2111: Fix duplicate key exception when a server drops and reconnects quickly
If a server drops and reconnects quickly then it might still have a valid session when it registers for a new session. The controller detects this by matching the callback URL and flags the old session to end while creating a new one.
The problem is that the old session might take a few ms to actually end and in this time we have two entries with the same calback url and hence the same serverIdentifier. If we try to get a server map in this time we will get a duplicate key exception.

The fix is to exclude servers from any processing once they have been flagged to end.

Related work items: #3745