Merged PR 2126: Fix for connection leak on getEntity().
Fix for connection leak on getEntity().
getEntity() was calling call.initialise an additional time to get the values, after it had already been initialised once to get the entity details. This was causing a new connection to be retrieved from the pool, resulting in the first call (for "getEntityDetails") to leak as the second ("getEntityValues") would now be the one being cleaned automatically. The fix was to enroll both calls in a transaction, which:
1. provides data consistency since we're now getting the values within the same transaction as the entity
2. cleanup the resources properly, and make it more readable and less error prone

Ultimately this would cause the connection pool to exhaust the connections, and when trying to load the value for the memory intent down the line, it would fail and simply return whatever WNET had returned (@meta.intent.XXX)

Related work items: #3891