Merged PR 2153: Return an error instead of the "@meta.intent.*" message if intent processing fails
We previously returned null internally, same as we would if there was no intent, so response processing would continue and the @meta tag would be returned to the user.
We now throw an exception internally if we began with an @meta but for any reason did not correctly match an intent. This leads to an error 500 going out to the user.

Related work items: #3905