Merged PR 2205: Prevented training from restarting when a skill is linked/unlinked since bot
Prevented training from restarting when a skill is linked/unlinked since bot linking only has any effect on chat, and not training.

Note that we will need to check if there is any edge case where the AI will not be accepting incoming chat requests because the training state won't be updated now.

Related work items: #4078