Merged PR 2206: Allow entities to be deleted if only in use by already deleted AIs.
Allow entities to be deleted if only in use by already deleted AIs.
Previously we weren't allowing entities to be deleted if there was any references to it, including already deleted AIs. This meant that once an entity was referenced by an AI and that AI was deleted, there was no way of deleting that entity anymore.

Related work items: #4035