Merged PR 2243: Re-introduce queue processor to AIML controller to allow db status updates.
Re-introduce queue processor to AIML controller to allow db status updates.
This was removed when the API split occurred since it doesn't have any other functional effect, but means that we stopped updating the controller status db table, meaning we couldn't track that and alert in case of going below a given threshold.

Related work items: #4104