Merged PR 2294: Send empty string on ApiAiDetails when training_file is null.
Send empty string on ApiAiDetails when training_file is null.
This is to support POJO-like implementations where the structure of the object is generated based on the received json and not mapped to a pre-defined class.

So when we have a=null, b=1, our json will look like {a=null, b=1} and our current GSON configuration will remove the nulls resulting in {b=1}. Now on the receiving side, if the type is constructed dynamically from the message, you end up with Class(b=1) with no references to "a".

Related work items: #4205