Merged PR 2379: Update chat logic to use a workflow approach instead of a monolithic static block of code
Update chat logic to use a workflow approach instead of a monolithic static block of code.
The chat workflow is defined by a sequence of chat handlers (IChatHandler). Each of the handlers is called in sequence until one of then (presumably one that has high confidence in the response) marks the chat as completed, and then all remaining handlers are skipped.
Currently the following handlers are available (and in the following order):
1. ChatPassthroughHandler - allows the chat to be completelly redirected to an external endpoint/hook
2. ChatIntentHander - processes any intents in flight
3. ChatRequestTrigger - this is an artificial handler, it's purpose is to send requests to the backends. We can later have separate request trigger handles per backend (for instance only sending to one backend if at that point in the workflow we don't have yet a valid response), but for now this follows current behaviour of triggering requests to all backends at once
4. ChatWnetHandler - handles any responses from WNET
5. ChatAimlHandler - handles any responses from AIML
6. ChatDefaultHandler - this is the final handler and always terminates the workflow. Basically it's the fallback whenever there are no responses found.

Related work items: #4317