Merged PR 2392: Added API support for the new SVM backend.
Added API support for the new SVM backend.
For now I've just added a new instance of controllers and connectors for SVM, we should consider making this more dynamic once we add more backend types.
The SVM backend for now is pretty much ignored by the chat flow, just processes it and logs it's results *if* it manages to process the request in less than one second (or whichever time remaining we have for completing the call, if less).

Related work items: #4281