Merged PR 2420: Update sort-by field on chat logs collection to dateTime.
Update sort-by field on chat logs collection to dateTime.
We've removed the timestamp field a long time ago, but it still existed in the index mappings. When we recreated the indices in DEV this was no longer available so it started failing the query to obtain the chatlogs. This problem would only occur in PROD if we'd recreate the indices, but better fix it now than have problems later.

Related work items: #4407