Merged PR 2620: Add support for 'last_updated' field on Intents.
Add support for 'last_updated' field on Intents.
This is now updated every time we update an intent.

The only downside is that the "last_updated" will get initialised to the date/time we run the script during deployment to PROD. To get this to reflect the "created_time" (which could potentially be incorrect as well) we'd need to create a trigger, which is overkill. Or run a one-off script after deployment to adjust the last_created to create_time...

Related work items: #4765