Merged PR 2664: Added auto-handover mechanism to allow the user to be handed over to another
Added auto-handover mechanism to allow the user to be handed over to another system after a sequence of low confidence responses from the AI.
This is an opt-in mechanism where the user will need to set up the new parameters to take advantage of this functionality, otherwise the behavious will remain the same as currently (the conversation is not handed over no matter how many low confidence responses are sent).
This introduces 3 new fields which can be set by the user:
- error_threshold_handover -> number of low confidence answers, in a row, that will trigger the automatic handover. If < 0 then no handover will happen
- handover_message -> message to send to the user when the conversation is being handed over. Note that this will only be sent once, after that the API will simply stop sending answers
- handover_reset_timeout -> the number of seconds while the handover will remain effective. After this time, the API will resume sending the requests to the backends and relay the responses to the user
Note that handover_message and handover_reset_timeout will only be used if error_threshold_handover > 0.

Related work items: #4751