Merged PR 2754: Update list json serialisation to account for type erasure.
Update list json serialisation to account for type erasure.
Basically when using generics, gson "loses" the type when trying to automatically determine the runtime type of the generic on .fromJson (this artifact is called Java Type Erasure). So to allow it to deserialise generic lists of objects (and not just built in types) one has to actually force the type at build-time (otherwise you end up with a LinkedTreeMap instead of the requested objet type).

Related work items: #4930