Merged PR 2808: Add intent condition evaluator and UTs.
Add intent condition evaluator and UTs.
This is just for evaluating a set of conditions, but later it will need to be wired up to gate triggering the intent.

Note that this PR is dependent on https://hutoma.visualstudio.com/Hutoma%20API/Hutoma%20API%20Team/_git/web-api/pullrequest/2807?_a=overview, so it will show more stuff until that one is checked in.

Related work items: #5041