Merged PR 2867: Added support for defining lifetime (number of turns) until a variable is
Added support for defining lifetime (number of turns) until a variable is discarded.
The value is for this lifetime can be any number of turns > 0. Any value provided during intent creation/update that is <= 0 will mean that the varible will not have a lifetime but instead exist for the whole session (or until the session is reset).

Related work items: #5124, #5137