Merged PR 2929: Added intent post-fulfillment conditional intent triggering logic.
Added intent post-fulfillment conditional intent triggering logic.
This allows an intent to be executed after another intent finishes, chaining intents. The new intent to be executed can be gated on a condition. In fact, there can be multiple intents attached to this step, and the first (defined) intent which fulfills all conditions will be executed (and any remaining will be ignored).
This supports having for instance:
- IF var=yes THEN EXEC intentYes
- IF var=no THEN EXEC intentNo
- (no condition) EXEC intentFallback  (this case this would be akeen to a "default" switch clause)

Cycle detection is performed to make sure you can't have A->B->C->A. To execute an intent it means all the conditions would have to be fulfilled, which means each intent would be executed with the same variable values (wrt the variables it cares about), so it would generate a true infinite loop.

Related work items: #5213