Merged PR 2974: Prevent retrieving entity values from what the user typed when we're in the
Prevent retrieving entity values from what the user typed when we're in the middle of an intent chain.
When an inner chained intent has an entity value that matches what the user typed (either as a intent trigger or prompt), this will automatically match on the inner intent and prevent the variable to be prompted.
Here is what was happening:
1. user types 'coffee'
2. intent is triggered, asks for the size
3. user types 'large'
4. intent coffee is fulfilled, triggers then intent cake
5. while we process the intent cake, we see it has a variable to fulfil
6. then first thing we do is to try to "extract" the value from what the user said (which works ok when this is the initially triggered intent). Since the last user question was "large", and our variable has it as a possible value, it auto-sets the cake.size to 'large' as well and (incorrectly) fulfils the intent.

This change prevents any entity extraction inside an intent chain.

Related work items: #5316