Merged PR 3112: Fix the training material encoding validation to use CP1252
Fix the training material encoding validation to use CP1252 (the one used by our current MySql installations) instead of ISO_8859_1 (which doesn't support certain variations of single and double quotes, and the Euro symbol, for instance).
Also now it doesn't fail on the first offending line, but will instead log all the offending lines so this can be more easily analysed later (in case there are characters we want to support but are not in the codepage).
Emoticons will still fail validation.

Related work items: #5541