Merged PR 3095: Updated intents storage to be written to/read from a single Json field.
Updated intents storage to be written to/read from a single Json field.
We only read/write an intent as a block so there's no need to have it in a more complex child-parent table relationship

Note: I've added an endpoint to support converting all the existing intents (from non-deleted AIs) into the new Json field to allow a one-time conversion. After that then the intent logic will only use the son field for reading and writing. Added @Deprecated to all the methods we can then remove after the conversion is done, so we can easily delete it later (and provide a visual queue to anyone who tries to use those methods)