Merged PR 3171: Added a feature toggle functionality to allow features to be toggled on/off.
Added a feature toggle functionality to allow features to be toggled on/off.
Features can be toggled globally, at a user-level and at an AIID level. An AIID setting will override a DEV setting which will override a Global setting.
The re-load of feature states is done from database at a given interval (when required) - default is every 60 seconds - but overridable in config - that feature states are being queried (note that no db access is done if no feature toggles are being queried).

Related work items: #5656