Merged PR 3182: Log webhook errors caused by external conditions.
Log webhook errors caused by external conditions.
Added an additional AppLog with common net failure reasons, so we can split the errors into buckets.
Changed the Passthrough external error from logging an exception to a warning, since this is not really an exception in the platform code but caused from an external call.
Used similar approach for passthrough and webhooks in terms of logging.

Related work items: #5673