Merged PR 3204: Fix linked skills when importing a bot in-place.
Fix linked skills when importing a bot in-place.
This uses the same approach as with intents, we first need to unlink all existing skills and then (re)link any skills that are in the imported bot.
Note that if the user doesn't own a skill on the imported bot, it will be ignored and only already owned skills will be linked (this is the default behavior for "normal" import as well)

Related work items: #5724