Merged PR 3055: Added a OWASP dependency scanner to the API build process.
Added a OWASP dependency scanner to the API build process.
This scans all included depdenencies in the jar for published vulnerabilities (CVEs) and aborts the build when at least one is found which is High (4).
Forced some dependencies to use the version withouth the CVE (as some libraries haven't updated yet their dependencies for this).
Removed dependency of google.common.Strings. This has known incompatible dependency issues around different versions which are pulled from different libraries, in particular with the latest versions we're now using (less or no CVEs), and replaced with the more stable Apache commons.

Related work items: #5745