Merged PR 3219: Prevent importing when conflicting entity values are found.
Prevent importing when conflicting entity values are found.
Since entities help defining the behavior of the intents, when importing entities with the same name as already existing ones, but with different values, we could be breaking intents. At the import stage we don't know if we should overwrite the existing or ignore the new ones, and since it's a stateless CURD operation we can't be going back and forth to the user to ask what to do on every conflict. So it's better just to fail the whole thing and let the user know that they have a conflict so they can go about making the decision.
An example would be, for instance, having an existing entity called rock (mineral). If I then import a bot that contains another entity called "rock" as well (music style), this is going to be conflicting. Deleting any of the new or existing would make their respective intents not work, and merging them would make both nonsensical.

Related work items: #5735