Merged PR 3301: Provide support for backend service registration defining language and version.
Provide support for backend service registration defining language and version.
A backend service (EMB at this point) can now provide the language it supports and which (BE) version it uses, to allow (later, not in this change) bots to use languages apart from EN, and to support experiments with new versions of backends.
Updated (devId, aiid) method signatures to use a new AiIdentity class so in the future we can further extend what identifies a bot without having to change the signatures of most methods (again).
Note that although the controller now manages different lang/version backends, the API sill still only respond to lang=EN and version=default requests (it will default to these values if not provided, to maintain the existing behaviours)

Related work items: #5828