Merged PR 3327: Small misc changes to explicitly indicate that our responses are UTF-8 and JSON.
Small misc changes to explicitly indicate that our responses are UTF-8 and JSON.
This doesn't affect the current behavior but makes it more consistent and compliant for different types of clients. For instance a curl call should now always get:
< Content-Encoding: UTF-8
< Content-Type: application/json
when receiving responses in JSON (which are pretty much all of them with the exception of chat logs download)

Also added (service) test cases for import/export to make sure UTF-8 content doesn't get changed when going through the Jersey stack.