Merged PR 3360: Fix entity deletion by validating if entity is in use
Fix entity deletion by validating if entity is in use

Until we provide bot-specific entities we need to go through all the existing intents and check if the entity we’re trying to delete is in use by any of them.
Note that we cannot just use server-side JSON functions on the db due to collation issues.

Related work items: #5937