Merged PR 3374: Disable fluent logger by default
Disable fluent logger by default
That fixes a load of debug warnings and at worst spurious crashes in
trying to connect to non-existing server.

Require an API_LOGGING_FLUENT_HOST to be set in config/environment
to enable it.

Checked that this is done in Docker already.

Related work items: #5963