Merged PR 3375: Add version_migrator service
Add version_migrator service
This will be accessible via HTTP interface
POST to retrain_all to retrain all bots that have training data
(in priority order)

GET to status to show what operation is ongoing and it's progress.

In docker so it logs as it goes along to ES.

The service talks to the DB as hutoma_caller, so it needs a SP.
to find bots to retrain.

The service talks to the API, using new endpoints under admin.
This is because we will be retraining bots
from potentially every single dev_id in the system
and we don't want to grab people's dev tokens.

Add it to the local docker-compose for testing.
Tested by using the stress test to create bots that need retraining.

First service that uses Python 3.7 which streamlines the
asyncio / aiohttp a bit.

Related work items: #5915