Merged PR 3398: Add getAllEndpoints endpoint to controller to support broadcast operations.
Add getAllEndpoints endpoint to controller to support broadcast operations.

Note that deleting a developer will send a request to all the registered endpoints to delete the data for that developer, but when deleting a bot only the currently configured language+version backend will be requested for deletion. This derives from deleting a DEV being a direct operation whereas deleting a bot is a queued one. Moving the delete into a direct operation would have more implications in terms of the current functionality and would require a much bigger change. So _if_ there are any orphaned bots, they will still be deleted upon deleting the developer.
We could at a later stage run a cleanup script to delete all the orphaned bots from NFS

Related work items: #5917