Merged PR 3468: Fix chat v2 protocol to support sending the ChatId as this is required by AIML.
Fix chat v2 protocol to support sending the ChatId as this is required by AIML.
AIML needs to determine which requests belong to the same chat session so that it can determine repetitions, and store state. So this needs to be passed on from the API to the backend service as previously (EMB will just ignore this, at least for now).
Also fixed a situation where a rogue entity (allowed to be inserted at one point and ot considered valid later) could cause a NPE during intent processing, so added some safeguards around that, and at least preventing the chat process from suffering a horrible death.

Related work items: #6072, #6100