Merged PR 3524: Fixes intent fulfilment by preventing variables used in previous calls to be reused
Fixes intent fulfilment by preventing variables used in previous calls to the same intent from living after the intent is fulfilled.
EntityValues are only supposed to exist during the lifetime of an intent, as are held to allow multiple intent variables to get prompted. At the end, they need to be cleared. With the introcduction of the ES entity values, these were no longer being correctedly tracked, and were now being persisted in the chat state even after the intent was completed (fulfilled or reset).

Related work items: #6162