Merged PR 3494: Prevent an error when updating the case of an existing entity value.
Prevent an error when updating the case of an existing entity value.
We were checking if the entity value existed by performing a case-sensitive comparison, and then hitting a constraint violation when trying to insert the value (as this is case-insensitive). This fix changes that comparison to be case-insensitive.
Note that changing letter case on the same word doesn't actually make it change in the db, as we handle entity values case-insensitively (if you have "AAA" and change it to "aaa" it will still be showing the original "AAA" - but when we use entity values we don't look at casing, so the results will be the same)

Related work items: #6115