Merged PR 3520: Add support for overriding the target engine version for the /admin/migration endpoints.
Add support for overriding the target engine version for the /admin/migration endpoints.
This allows migrating a bot to a new engine version without downtime. The idea is to:
1. call POST /admin/migration/{devid}/{aiid}?engine_version={new version}
this schedules training for the given bot on the backend with the given version. The existing bot is still fully available for training and chatting as it's being served by the backends with the previous version. As far as the clients (UI, direct API) are concerned, nothing has changed
2. poll until GET /admin/migration/{devid}/{aiid}?engine_version={new version} returns that the training has completed.
3. call PUT /admin/migration/{devid}/{aiid}/version?engine_version={new version} to switch the backend version that serves the bot.
from now on, the new backend version is live and serving training and chatting requests.
Note that step 3 can be called after the nre training has completed for each bot, or alternatively it can be called for all bots at the end. It doesn't really matter (for the user) as they experience no downtime, but it provides options for moving users onto new versions and not having to commit to it once the retraining process starts

If any of the admin/migration endpoints is called without the "engine_version" param, then it will operate on the current version (legacy behaviour) which will take each bot offline while it's training.

(this is dependent on PR 3511)

Related work items: #6215