Merged PR 3638: Explicitly require DevId validation on endpoints using DevId
Explicitly require DevId validation on endpoints using DevId

We were always expecting a DevID to be presented in an authentication header, which caused some issues during validation (DevID was validated as long as there was a ValidateXXX attribute in the endpoint, but no validation was performed if there was no other param to validate). This now consistently requires parameters to be validated (in particularly DevID) to be explicitly declared.
This now also allows for having endpoints which don't use the Bearer authentication scheme with a JWT including a DevID, allowing other mechanisms to be used as well.

Related work items: #6403