Merged PR 3642: Per Bot Entities Part 1: Updates that keep everything the same
This is the first of 2 updates for per bot entities. This makes a few db changes that let everything operate the same as it does now. Once this works we can put the updates that pass the aiid in under a feature toggle so that we can flip back and forth initially.

Related work items: #6203, #6205, #6206