Merged PR 3655: Add webhook callback endpoint
Add webhook callback endpoint

The webhook callback endpoint allows a webhook to inject a message into the chat flow (currently only for Facebook). This effectively allows a webhook to asynchronously provide feedback about a long running/UI operation. This also allows the webhook to set chat context variables to update the state.

A call to this endpoint looks something like:
```
curl -k -X POST \
-H "Content-type: application/json" --data '{"webhook_token":"eyJhbGciOiJIUzI1NiIsInppcCI6IkRFRiJ9.eNoczjsOgjEMA-C7dCZS0qaPsCFY_p0LpG0iJCQYgAlxdwqTl8-W3-GwbaewD7N2bSQd2DUB91xAsBlERlMVzyvDLpzvV7ttcxWc0JoIwcgegduyomLgWilOjeLVV-F40effk3pkmgnMbUAqqUPjQZAQybtE71iXf7z6wsprHUcDscHAWDLo7012qxhN5-QcPl8AAAD__w.E-a4szKXQU4F4Dy37FxXHq7ewCLPN3E3yjoGC-kWOA4", "text":"The text to be injected"}' \
https://<API SERVER>/v1/ai/chat/webhook/callback/73820abc8668fe7e0fc3a30ac9957b345f02ff3a7bb3967ec0c6b585debbd21b
```
The path parameter after `callback` is the SHA256'd chatId, and the `webhook_token` field is the JWT token generated before the webhook is called and passed on to it.

Related work items: #6398, #6399