Merged PR 4011: Allow usage of sys.any inline with the user’s question
Allow usage of sys.any inline with the user’s question

SysAnyCommands allows to intercept the conversation flow in search for an intent pattern like
trigger_command @{sys.any}@
These are typically sent by automated processes (typically webhooks) to trigger an action in the conversation, and require it to be precise instead of going through the NLP processing. For instance situations where the input flow is constructed artificially like:
ADD_PRODUCT_TO_CART PROD_ID_1234
we cannot rely on the NLP to try to extract a conversational intent from it, but we need to trigger the appropriate intent to allow the external process to move forward.
Note that we require explicitly that the intent trigger follows a very rigid format, otherwise we will fallback to the default chat behaviour.
This behaviour will be shipped initially behind a feature toggle (SYSANY_COMMANDS)

Also:
- changed the chat pipeline to only call the ER if we’re not circumventing the chat process
- updated the Elasticsearch libraries to a version which currently is CVE clean

Related work items: #7358